Fixed removing filtered words

+++ utilities/autowarning.js
@@ -24,18 +24,22 @@ module.exports.globalfilter = (client, message, sqlcon) => {
     })
 }
 module.exports.filter = (client, message, sqlcon) => {
-    sqlcon.query(`SELECT * FROM filter WHERE GuildID = '${message.guild.id}'`, (err, filtered) => {
-        if (filtered != undefined) {
-            filtered.forEach(element => {
-                if (message.content.includes(element.Word)) {
-                    let AutoWarnReason = "Saying a filtered word"
-                    let AWUser = message.author
-                    let AWMember = message.member
-                    let issueTime = moment(Date.now()).format('DD MMM YYYY, HH:mm')
-                    AddAutoWarn(client, AutoWarnReason, AWUser, AWMember, issueTime, sqlcon, message)
-                    message.delete().catch(error => { console.log(error) })
+    sqlcon.query(`SELECT * FROM guildprefs WHERE GuildID = '${message.guild.id}'`, (err, rows) => {
+        if (!message.content.includes("filter remove") && !message.member.hasPermission("ADMINISTRATOR") && !message.member.roles.find(role => role.id === rows[0].AdminRole)) {
+            sqlcon.query(`SELECT * FROM filter WHERE GuildID = '${message.guild.id}'`, (err, filtered) => {
+                if (filtered != undefined) {
+                    filtered.forEach(element => {
+                        if (message.content.includes(element.Word)) {
+                            let AutoWarnReason = "Saying a filtered word"
+                            let AWUser = message.author
+                            let AWMember = message.member
+                            let issueTime = moment(Date.now()).format('DD MMM YYYY, HH:mm')
+                            AddAutoWarn(client, AutoWarnReason, AWUser, AWMember, issueTime, sqlcon, message)
+                            message.delete().catch(error => { console.log(error) })
+                        }
+                    });
                 }
-            });
+            })
         }
     })
 }
